requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.0
filelock==3.12.4
fonttools==4.44.0
fsspec==2023.10.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jose==1.0.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
mysql-connector-python==8.1.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
passlib==1.7.4
Pillow==10.1.0
platformdirs==3.11.0
protobuf==4.21.12
psutil==5.9.6
py-cpuinfo==9.0.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.208
urllib3==2.0.7
uvicorn==0.23.2
virtualenv==20.24.5
